requirements.txt
flask 
python-dotenv 
requests
replicate
openai
sqlite3